requirements.txt
matplotlib==3.8.3
networkx==3.2.1
numpy==2.1.3
openai==1.55.0
PyYAML==6.0.2
seaborn==0.13.2